requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pymupdf==1.24.10
python-docx==1.1.2
fastapi==0.112.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2